Add front end side of MERN tutorial
Add front end side of MERN tutorial from Jelo Rivera, "Let's build a
full stack MongoDB, React, Node and Express (MERN) app" on Medium.com.
This is not yet finished based on the tutorial, more will be added after
the backend is set up, and again to modify the files for Room-Arranger.